+++ src/main/java/net/pointofviews/movie/service/impl/MovieTMDbSearchService.java
@@ -1,7 +1,8 @@
-package net.pointofviews.movie.service;
+package net.pointofviews.movie.service.impl;
 
 import net.pointofviews.movie.dto.response.SearchMovieApiListResponse;
 import net.pointofviews.movie.exception.MovieException;
+import net.pointofviews.movie.service.MovieApiSearchService;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpRequest;
