+++ src/test/java/net/pointofviews/movie/service/MovieTMDbSearchServiceTest.java
@@ -5,6 +5,7 @@ import lombok.SneakyThrows;
 import net.pointofviews.movie.dto.response.SearchMovieApiListResponse;
 import net.pointofviews.movie.dto.response.SearchMovieApiResponse;
 import net.pointofviews.movie.exception.MovieException;
+import net.pointofviews.movie.service.impl.MovieTMDbSearchService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
